deployment/requirements.txt
torch==1.13.1
sanic==21.9.3
sanic_cors==2.0.0
nltk==3.7
numpy==1.23.3
transformers==4.23.1
typing==3.7.4.3
lrparsing~=1.0.13
bson==0.5.10
python-dotenv==0.21.0
openai==0.26.0
zhon==1.1.5
iso-639==0.4.5
